+++ final_task/rss_reader/RSSReaderException.py
@@ -1,0 +1,2 @@
+class RSSReaderException(Exception):
+    pass
